requirements.txt
scrapy
torch
clip
torchvision
streamlit
qdrant-client
boto3
Pillow